Add CLI and API specs to search results
