+++ apps/docs/scripts/build-search.ts
@@ -1,84 +1,21 @@
-const fs = require('fs')
-const crypto = require('crypto')
-const path = require('path')
-const matter = require('gray-matter')
-const dotenv = require('dotenv')
-const algoliasearch = require('algoliasearch/lite')
-const { isEmpty } = require('lodash')
-const yaml = require('js-yaml')
-const commonLibSections = require('../../../spec/common-client-libs-sections.json')
+import fs from 'fs'
+import crypto from 'crypto'
+import path from 'path'
+import matter from 'gray-matter'
+import dotenv from 'dotenv'
+import algoliasearch from 'algoliasearch/lite'
+import { isEmpty } from 'lodash'
 
-function flattenSections(sections) {
-  var a = []
-  for (var i = 0; i < sections.length; i++) {
-    if (sections[i].id) {
-      // only push a section that has an id
-      // these are reserved for sidebar subtitles
-      a.push(sections[i])
-    }
-    if (sections[i].items) {
-      // if there are subitems, loop through
-      a = a.concat(flattenSections(sections[i].items))
-    }
-  }
-  return a
-}
+// search objects
+import { generateClientLibSearchObjects } from './files/client-libs'
+import { generateAPISearchObjects } from './files/api'
+import { generateCLISearchObjects } from './files/cli'
 
-const clientLibFiles = [
-  { fileName: 'supabase_js_v2', label: 'javascript', version: 'v2', versionSlug: false },
-  { fileName: 'supabase_js_v1', label: 'javascript', version: 'v1', versionSlug: true },
-  { fileName: 'supabase_dart_v1', label: 'dart', version: 'v1', versionSlug: false },
-  { fileName: 'supabase_dart_v0', label: 'dart', version: 'v0', versionSlug: true },
-]
+const cliObjects = generateCLISearchObjects()
+const apiObjects = generateAPISearchObjects()
+const clientLibSearchObjects = generateClientLibSearchObjects()
 
-const nameMap = {
-  api: 'Management API',
-  cli: 'Supabase CLI',
-  auth: 'Auth Server',
-  storage: 'Storage Server',
-  postgres: 'Postgres',
-  dart: 'Supabase Flutter Library',
-  javascript: 'Supabase JavaScript Library',
-}
-
-const flatCommonLibSections = flattenSections(commonLibSections)
-
-// loop through each spec file, find the correspending entry in the common file and grab the title / description / slug
-let clientLibSearchObjects = []
-
-clientLibFiles.map((file) => {
-  const specs = yaml.load(fs.readFileSync(`../../spec/${file.fileName}.yml`, 'utf8'))
-
-  //take each function id, find it in the commonLibSections file and return { id, title, slug, description, }
-  specs.functions.map((fn) => {
-    const item = flatCommonLibSections.find((section) => section.id === fn.id)
-    if (item) {
-      const object = clientLibSearchObjects.push({
-        objectID: crypto.randomUUID(),
-        id: item.id,
-        title: item.title,
-        description: item.title,
-        url: `/reference/${file.label}/${file.versionSlug ? file.version + '/' : ''}${item.slug}`,
-        source: 'reference',
-        pageContent: '',
-        category: item.product,
-        version: file.version,
-        type: 'lvl2',
-        hierarchy: {
-          lvl0: 'References',
-          lvl1: `${nameMap[file.label]} ${file.version}`,
-          lvl2: item.title,
-          lvl3: file.version,
-          lvl4: null,
-          lvl5: null,
-          lvl6: null,
-        },
-      })
-      return object
-    }
-  })
-})
-
+// @ts-ignore
 // The properties of the searchObject that are specifically read
 // by DocSearch are "type" and "hierarchy". The rest, though saved into Algolia (which we
 // can potentially use to craft more nuanced search experiences) are not used by DocSearch
@@ -97,6 +34,7 @@ const ignoredFiles = [
 
 async function walk(dir) {
   let files = await fs.promises.readdir(dir)
+  //@ts-ignore
   files = await Promise.all(
     files.map(async (file) => {
       const filePath = path.join(dir, file)
@@ -131,7 +69,7 @@ async function walk(dir) {
 
     const guidePages = (await walk('pages')).filter((slug) => !ignoredFiles.includes(slug))
 
-    // generate search objects for guide pages
+    // generate search objects for mdx guide pages
     const guidePagesearchObjects = guidePages
       .map((slug) => {
         let id, title, description
@@ -196,13 +134,20 @@ async function walk(dir) {
       .filter((object) => !object.url.endsWith('/index'))
       .filter((object) => !object.url.endsWith('/.gitkeep'))
 
-    const combinedSearchObjects = guidePagesearchObjects.concat(clientLibSearchObjects)
+    const combinedSearchObjects = guidePagesearchObjects.concat(
+      clientLibSearchObjects,
+      apiObjects,
+      cliObjects
+    )
 
+    //@ts-ignore
     await index.clearObjects()
     console.log(`Successfully cleared records from ${indexName}`)
 
+    //@ts-ignore
     const algoliaResponse = await index.saveObjects(combinedSearchObjects)
 
+    //@ts-ignore
     console.log(`Successfully saved ${algoliaResponse.objectIDs.length} records into ${indexName}.`)
   } catch (error) {
     console.log('Error:', error)
