+++ apps/docs/scripts/helpers.ts
@@ -1,0 +1,9 @@
+export const nameMap = {
+  api: 'Management API',
+  cli: 'Supabase CLI',
+  auth: 'Auth Server',
+  storage: 'Storage Server',
+  postgres: 'Postgres',
+  dart: 'Supabase Flutter Library',
+  javascript: 'Supabase JavaScript Library',
+}
